# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking crypto portfolios, transactions, and prices",
        "title": "Crypto Portfolio Tracker API",
        "contact": {
            "name": "David Nwaekwu",
            "email": "nwaekwudavid@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/prices": {
            "post": {
                "description": "Fetch USD prices for tokens by chain + contract address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prices"
                ],
                "summary": "Get token prices",
                "parameters": [
                    {
                        "description": "Assets to price",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PricesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PriceAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallets/{wallet}/portfolio": {
            "get": {
                "description": "Fetch wallet portfolio with live valuation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolio"
                ],
                "summary": "Get portfolio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PortfolioResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallets/{wallet}/portfolio/holdings": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolio"
                ],
                "summary": "Update holding",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Holding",
                        "name": "holding",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/portfolio.Holding"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolio"
                ],
                "summary": "Add holding",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Holding",
                        "name": "holding",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/portfolio.Holding"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolio"
                ],
                "summary": "Delete holding",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Holding",
                        "name": "holding",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/portfolio.Holding"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/wallets/{wallet}/transactions": {
            "get": {
                "description": "Fetch paginated transactions for a wallet",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "List wallet transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blockchain (ethereum)",
                        "name": "chain",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Token symbol",
                        "name": "token",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date RFC3339",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date RFC3339",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TransactionListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.AssetRequest": {
            "type": "object",
            "properties": {
                "chain": {
                    "type": "string"
                },
                "contract_address": {
                    "type": "string"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.PortfolioResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/portfolio.PortfolioView"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "description": "human-readable",
                    "type": "string"
                },
                "status": {
                    "description": "success | error",
                    "type": "string"
                }
            }
        },
        "handlers.PriceAPIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.PricesResponse"
                },
                "error": {
                    "$ref": "#/definitions/handlers.ErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "handlers.PricesRequest": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.AssetRequest"
                    }
                }
            }
        },
        "handlers.PricesResponse": {
            "type": "object",
            "properties": {
                "prices": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                }
            }
        },
        "handlers.TransactionListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "items": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/transactions.Transaction"
                            }
                        },
                        "limit": {
                            "type": "integer"
                        },
                        "page": {
                            "type": "integer"
                        }
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "portfolio.Holding": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "the amount owned",
                    "type": "number",
                    "format": "float64"
                },
                "chain": {
                    "description": "ethereum, polygon, etc",
                    "type": "string"
                },
                "contractAddress": {
                    "description": "empty for native asset",
                    "type": "string"
                }
            }
        },
        "portfolio.HoldingView": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "format": "float64"
                },
                "chain": {
                    "type": "string"
                },
                "contractAddress": {
                    "type": "string"
                },
                "priceUSD": {
                    "type": "number",
                    "format": "float64"
                },
                "valueUSD": {
                    "type": "number",
                    "format": "float64"
                }
            }
        },
        "portfolio.PortfolioView": {
            "type": "object",
            "properties": {
                "holdings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/portfolio.HoldingView"
                    }
                },
                "totalValueUSD": {
                    "type": "number",
                    "format": "float64"
                },
                "wallet": {
                    "type": "string"
                }
            }
        },
        "transactions.Direction": {
            "type": "string",
            "enum": [
                "in",
                "out"
            ],
            "x-enum-varnames": [
                "DirectionIn",
                "DirectionOut"
            ]
        },
        "transactions.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "format": "float64"
                },
                "chain": {
                    "description": "ethereum, polygon, etc",
                    "type": "string"
                },
                "direction": {
                    "$ref": "#/definitions/transactions.Direction"
                },
                "from": {
                    "type": "string"
                },
                "hash": {
                    "description": "on-chain tx hash",
                    "type": "string"
                },
                "id": {
                    "description": "internal ID (can be tx hash)",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/transactions.TransactionStatus"
                },
                "timestamp": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "token": {
                    "description": "ETH, USDC, WBTC, etc",
                    "type": "string"
                },
                "tokenAddr": {
                    "description": "empty for native ETH",
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/transactions.TransactionType"
                }
            }
        },
        "transactions.TransactionStatus": {
            "type": "string",
            "enum": [
                "success",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusSuccess",
                "StatusFailed"
            ]
        },
        "transactions.TransactionType": {
            "type": "string",
            "enum": [
                "send",
                "receive",
                "swap",
                "stake"
            ],
            "x-enum-varnames": [
                "TypeSend",
                "TypeReceive",
                "TypeSwap",
                "TypeStake"
            ]
        }
    }
}